\name{interMutDist}
\alias{interMutDist}
\title{
Calculate The Inter-mutational Distances
}
\description{
The function will calculate the inter-mutational distances after the ordering of
the variants according to the genomic positions.
}
\usage{
interMutDist(data)
}
\arguments{
  \item{data}{
the read-in matrix or list of matrix generated by the readVCF or readMAF. See \
'Details'.
}
}
\details{
All the variants in a single matrix or a list of matrixes will be sorted
according to the genomics postition. Then the distance between two variants will
be calculated.
}
\value{
A dataframe will be returned with the following columns:
  \item{CHR }{The chromosome of the variant}
  \item{POS }{The position of the variant}
  \item{DISTANCE }{The distance of the variant to its neighbour variant}
}
\references{
To be published
}
\author{
Xue Lin, Jian Li
}
\note{
If there is too little variants for each chromsome to calculate the
inter-mutational distances, a warning will rise, and the program will stop.
}
\seealso{
readVCF, readMAF
}
\examples{
#library(kataegis)

##Read in MAF file and merge all the samples in one file:
fpath <- system.file("extdata", "test_mut.maf", package="kataegis")
dat<-readMAF(fpath)

#Read in MAF file and split the maf according the samples' ID:
list_dat<-readMAF(fpath, split=TRUE)

#Or you can also import maf file typing
#dat<-readMAF("path/to/your/file.maf")
#lsit_dat<-readMAF("path/to/your/file.maf", split=TRUE)

##Read in VCF file:
fpath <- system.file("extdata", "test_mut.vcf", package="kataegis")
dat<-readVCF(fpath)

#Or you can also import vcf file typing
#dat<-readVCF("path/to/your/file.vcf")

##calculate the inter-mutational distances
dist<-interMutDist(dat)
#list_dist<-interMutDist(list_dat)
}
\keyword{ math }% use one of  RShowDoc("KEYWORDS")
